+++ Examples/Examples/ExamplesDefaults.swift
@@ -39,6 +39,7 @@ struct ExamplesDefaults {
         chartSettings.axisStrokeWidth = 1
         chartSettings.spacingBetweenAxesX = 15
         chartSettings.spacingBetweenAxesY = 15
+        chartSettings.labelsSpacing = 0
         return chartSettings
     }
     
@@ -54,6 +55,7 @@ struct ExamplesDefaults {
         chartSettings.axisStrokeWidth = 0.2
         chartSettings.spacingBetweenAxesX = 8
         chartSettings.spacingBetweenAxesY = 8
+        chartSettings.labelsSpacing = 0
         return chartSettings
     }