+++ Examples/Examples/StackedBarsExample.swift
@@ -70,7 +70,7 @@ class StackedBarsExample: UIViewController {
         let coordsSpace = ChartCoordsSpaceLeftBottomSingleAxis(chartSettings: chartSettings, chartFrame: chartFrame, xModel: xModel, yModel: yModel)
         let (xAxisLayer, yAxisLayer, innerFrame) = (coordsSpace.xAxisLayer, coordsSpace.yAxisLayer, coordsSpace.chartInnerFrame)
         
-        let barViewSettings = ChartBarViewSettings(animDuration: 0.5)
+        let barViewSettings = ChartBarViewSettings(animDuration: 0.5, cornerRadius: 6)
         let chartStackedBarsLayer = ChartStackedBarsLayer(xAxis: xAxisLayer.axis, yAxis: yAxisLayer.axis, innerFrame: innerFrame, barModels: barModels, horizontal: horizontal, barWidth: 40, settings: barViewSettings, stackFrameSelectionViewUpdater: ChartViewSelectorAlpha(selectedAlpha: 1, deselectedAlpha: alpha)) {tappedBar in
             let chartViewPoint = tappedBar.layer.contentToGlobalCoordinates(CGPoint(x: tappedBar.barView.frame.midX, y: tappedBar.stackedItemViewFrameRelativeToBarParent.minY))!
             let viewPoint = CGPoint(x: chartViewPoint.x, y: chartViewPoint.y + 70)