+++ SwiftCharts/Axis/ChartAxisYLowLayerDefault.swift
@@ -98,4 +98,8 @@ class ChartAxisYLowLayerDefault: ChartAxisYLayerDefault {
         }
         return labelsX
     }
+    
+    override func axisLineX(offset: CGFloat) -> CGFloat {
+        return self.offset + offset - settings.axisStrokeWidth / 2
+    }
 }